requirements.txt
# Core Dependencies / Основные Зависимости
numpy>=1.21.0
scipy>=1.7.0
pyyaml>=5.4.0

# Machine Learning / Машинное Обучение
torch>=1.9.0

# Visualization / Визуализация
matplotlib>=3.4.0
seaborn>=0.11.0

# Testing / Тестирование
pytest>=6.2.0
pytest-cov>=2.12.0

# Development / Разработка
black>=21.0
flake8>=3.9.0
mypy>=0.910
pre-commit>=2.13.0
isort>=5.10.0

# Documentation / Документация
mkdocs>=1.2.0
mkdocs-material>=7.0.0

# Optional: LLM Experiments / Опционально: LLM Эксперименты
# openai>=0.27.0
# transformers>=4.20.0